Build a static library and use it for main and the tests
With this, a static library is created from all the source files (except
main.cpp). Then the final binary is created by compiling main.cpp and
linking it against the library.

This allows us to link the tests.cpp also against the same library, without
having to list the source files again.