Make all IndexParameters attributes const
Also, add index parameters to CommandLineOptions struct and fill in those
instead of creating an IndexParameters object when parsing the command
line parameters. This make sense because the command-line parser doesn’t
need to know which parameters are used for which purpose.

Instead of a adjust_index_params_depending_on_read_length, we now have a
static method (this is the "named constructor idiom") that creates an
IndexParameters instance that doesn’t need to be modified anymore.
Immutable objects are nice because they are easier to reason about.