Get rid of IndexCreationStatistics::elapsed_copy_flat_vector
This measures "time copying flat vector".

It is the only duration that is "nested" twice: elapsed_flat_vector includes
elapsed_copy_flat_vector, which is somewhat confusing in the output because
numbers don’t add up to the total indexing time. Also the times for
elapsed_flat_vector are nearly identical to elapsed_copy_flat_vector anyway
(the only extra relevant thing that is measured is count_unique_elements,
which is fast).