Speed up uppercasing the reference
Reading a 3 GB reference takes about 1.8 seconds, but converting it to
uppercase with std::toupper takes another 6 seconds because toupper takes
the current locale into account. Since we assume ASCII for the FASTA input
file anyway, we can use a custom function that manipulates the bits of the
char. This takes only ~0.2 seconds.