Fix ReferenceMer memory usage
I added "bit_alloc" and "mask" as attributes in a previous commit, but did
not realize that a "const" attribute is still stored in the instance. It
needs to be static to avoid this.

I verified that the compiled code is the same as if one had hardcoded the
constants.