Show a progress indicator
During mapping, the progress indicator displays the number of mapped
reads and the (wall-clock) time per read.

The progress indicator is not shown within the first second, which ensures
that we don’t have misleading numbers on the screen in case only a small
number of reads were mapped.

Also, the progress indicator is only shown if the program is run
interactively (when stderr is an actual terminal) to avoid cluttering log
files when the user redirects stderr.