Support reading gzip-compressed references
We use the zstr external library for this. It wraps zlib in a C++-compatible
way so that we can use the FASTA parsing function unchanged.

While zstr can also read uncompressed input, it is a bit slower doing so
than when using a regular std::ifstream. To avoid a performance regression,
we use zstr::ifstream only if the input file name ends with .gz but fall
back to std::ifstream otherwise.

Supersedes and closes #231