Keep forward and reverse hits separate
merge_hits_into_nams() has a nested for loop that iterates over a vector
that contains both forward and reverse hits, but it also has an extra check
to ensure that it never merges forward and reverse hits into NAMs. We can
instead keep forward and reverse hits separate when we generate them and
then work on them separately.