Replace xxh64 with a simplified version
This function gives the same result as regular xxh64(), but the difference
is that it only works on single 64-bit values.

Mostly the "bit mixing" part of xxh64 (functions finalize()/avalanche()) is
left.

This performs a little bit better than calling the original function on an
8-byte slice because the compiler does not fully inline everything
(finalize() remains a function call).

This speeds up index generation by a couple of percent.