Filter by reversed randstrobe to recover lost accuracy
We lost some accuracy when switching to asymmetric randstrobes, and with this
commit, we recover all of it. That is, use of asymmetric randstrobes becomes
as accurate as using symmetric randstrobes.

The lost accuracy was due to filtering working differently: With symmetric
randstrobes, the decision of whether to filter a hit or not was based on how
often the randstrobe *or its reversed version* occur in the reference (which
is the desired behavior). With randstrobes becoming asymmetric, the decision
became based on only how often the forward version occurs. This leads to a
directional bias and a significant loss in accuracy (very apparent on the
highly repetitive chrY of CHM13).

Here, we restore the old, less biased filtering behavior by explicitly
adding up the number of occurrences of the forward and reversed randstrobe
and basing the filtering decision on that total count.

This involves some extra hash computations and additional index lookups, but
overall, strobealign actually becomes faster (up to 30% of chrY for certain
read lengths, more moderate improvements for the other references), likely
because fewer "spurious" hits are produced

Is-new-baseline: yes